requirements.txt
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
cryptography>=42.0
PyJWT>=2.8
javalang>=0.13
